requirements.txt
PyMuPDF>=1.23.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
beautifulsoup4>=4.12.0
flask>=3.0.0
